+++ src/JsonApi/Exception/JsonApiExceptionInterface.php
@@ -4,7 +4,7 @@ namespace WoohooLabs\Yin\JsonApi\Exception;
 interface JsonApiExceptionInterface
 {
     /**
-     * @return \WoohooLabs\Yin\JsonApi\Transformer\AbstractErrorDocument
+     * @return \WoohooLabs\Yin\JsonApi\Document\AbstractErrorDocument
      */
     public function getErrorDocument();
 }